requirements.txt
opencv-python==4.7.0.72
SpeechRecognition==3.8.1
PyAudio==0.2.11
pyttsx3==2.90
mediapipe==0.10.21
numpy==1.23.5
PyYAML==6.0
ultralytics==8.1.1
torch==2.2.0
torchvision==0.17.0